requirements.txt
alembic==1.5.2
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
distro==1.5.0
dnspython==2.1.0
docker==5.0.0
docker-compose==1.29.0
dockerpty==0.4.1
docopt==0.6.2
dominate==2.6.0
elasticsearch==7.12.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
paramiko==2.7.2
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
texttable==1.6.3
urllib3==1.26.4
visitor==0.1.3
websocket-client==0.58.0
Werkzeug==1.0.1
WTForms==2.3.3
